requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
PyPDF2
python-dotenv
google-cloud-storage
ipykernel
fastapi[standard]
uvicorn[standard]
python-multipart